refactor: merge mybatis pg/oracle updateInterceptor config test
