+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/config/MapperConfig.java
@@ -73,7 +73,7 @@ public class MapperConfig {
      */
     @Bean
     @ConditionalOnExpression(value = "'${shenyu.database.dialect}' == 'oracle' || '${shenyu.database.dialect}' == 'postgresql'")
-    public SqlUpdateInterceptor oracleSqlUpdateInterceptor() {
+    public SqlUpdateInterceptor sqlUpdateInterceptor() {
         return new SqlUpdateInterceptor();
     }
