+++ shenyu-admin/src/test/java/org/apache/shenyu/admin/config/MapperConfigTest.java
@@ -51,12 +51,7 @@ public class MapperConfigTest {
     }
 
     @Test
-    public void oracleSqlUpdateInterceptorTest() {
-        assertNotNull(mapperConfig.oracleSqlUpdateInterceptor());
-    }
-
-    @Test
-    public void postgreSqlUpdateInterceptorTest() {
-        assertNotNull(mapperConfig.postgreSqlUpdateInterceptor());
+    public void sqlUpdateInterceptorTest() {
+        assertNotNull(mapperConfig.sqlUpdateInterceptor());
     }
 }
